feat(tests): implement working PostgreSQL test setup for baseline tests
  - Add comprehensive Docker-based PostgreSQL container management
  - Implement automatic database schema initialization
  - Fix type compatibility issues (i32 to i64 for finding_id)
  - Create test project to satisfy foreign key constraints
  - Add robust container lifecycle management with unique naming
  - Ensure all 11 baseline tests pass successfully

  The baseline tests now have a complete, working PostgreSQL test environment
  that automatically manages containers and database setup, making the tests
  reliable and maintainable.

+++ crates/hodei-server/src/modules/types.rs
@@ -195,7 +195,7 @@ pub struct Project {
 /// Baseline status for a finding
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BaselineStatus {
-    pub finding_id: i32,
+    pub finding_id: i64,
     pub status: FindingStatus,
     pub reason: Option<String>,
     pub expires_at: Option<DateTime<Utc>>,
